Refactor and improve codebase
