+++ eventLoader.js
@@ -1,0 +1,13 @@
+const { readdir } = require('fs');
+
+module.exports = async (client) => {
+  readdir('./events/', (err, files) => {
+    if (err) return console.error(err);
+
+    files.forEach((file) => {
+      const event = require(`./events/${file}`);
+      let eventName = file.split('.')[0];
+      client.on(eventName, event.bind(null, client));
+    });
+  });
+};
