+++ example.env
@@ -1,0 +1,2 @@
+TOKEN=
+PREFIX=
