+++ package.json
@@ -9,8 +9,10 @@
   "author": "Salty-Coder",
   "license": "ISC",
   "dependencies": {
+    "chalk-advanced": "^1.0.6",
     "cheerio": "^1.0.0-rc.3",
     "discord.js": "^14.5.0",
+    "dotenv": "^16.0.3",
     "request": "^2.88.2"
   }
 }
